+++ src/contracts/BaseCommand.ts
@@ -55,9 +55,7 @@ export abstract class BaseCommand implements Command {
      * @description Execution context
      */
     public readonly context: Context,
-  ) {
-    this.synchronize()
-  }
+  ) {}
 
   /// region abstract methods
   /**
@@ -73,15 +71,6 @@ export abstract class BaseCommand implements Command {
    * @return {string}
    **/
   public abstract get descriptor(): string
-
-  /**
-   * Synchronize the command execution with the network. This method shall
-   * be used to fetch data required for execution.
-   *
-   * @async
-   * @return {Promise<boolean>}
-   */
-  public abstract async synchronize(): Promise<boolean>
 
   /**
    * Verifies **allowance** of `actor` to execute command.  Arguments to